Unhighlight the htmlapi-client-python project.
This project was demo code that went along with a conference talk
given by Jon Moore and was released for reference by request of
several folks in the audience. However, it is really just a body
of reference code and not an active project. The plan is to
migrate this over to a public repository under @comcast-jonm,
which is a more appropriate location for it.